requirements.txt
absl-py==0.14.1
antlr4-python3-runtime==4.8
anyio==3.3.0
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cloudpickle==1.6.0
cycler==0.10.0
dask==2021.8.1
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dm-tree==0.1.6
entrypoints==0.3
flatbuffers==2.0
fsspec==2021.8.1
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.17
greenlet==1.1.2
grpcio==1.41.0
h5py==3.1.0
hjson==3.0.2
idna==3.2
iniconfig==1.1.1
install==1.3.4
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
llvmlite==0.36.0
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numba==0.53.0
numpy==1.22.2
oauthlib==3.1.1
opendp==0.3.0
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandasql==0.7.3
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.6.0
pluggy==1.0.0
preprocess==2.0.0
prettytable==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pynndescent==0.5.5
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-allel==1.3.5
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.31
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-probability==0.15.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==2.2.0
tomli==2.0.0
toolz==0.11.1
torch==1.10.2
torchvision==0.8.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.7.4.3
umap-learn==0.5.2
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
